launch-softlayer: fix reboot and upgrade, flake8, add wait_for_ssh
Modified runcmd some so that caller can pass in timeout and see rc.
Added wait_for_ssh that will block waiting on the ssh to be up.
This was needed after reboot.
Change wait_on_cloud_init to use cloud-init status --wait.
Improve proposed to not require getting the distro from the platform,
but rather read it from the system.

The overall change here is that --proposed and --clean now works as
expected:
  launch-softlayer --proposed --keep-alive --clean -v --series=xenial