lxc-proposed-snapshot: return full path to commands in find_command
find_command() is used to determine the path to a command, which is then
used to push that command in to the lxd container being produced.
Currently, if the command is on the user's PATH, instead of returning
the full path, it will return just the name of the command.  As such,
this only works if you're already in the directory containing the
command.

This changes find_command to instead return the full path, fixing
lxc-proposed-snapshot invocation when you have update-root on your PATH.