pstart: Do not re-apply possibly stale config.
Previously in cleanup, pstart would set a container's config
back to the cached value.  That caused problems because of the
volatile entries (https://github.com/lxc/lxd/issues/6632).

Instead, use 'lxc profile add' and 'lxc profile remove' to add
and remove the profile.  Unfortunately we have to use 'lxc config'
to list the current profiles of the container.  There is no lxc
profile container-list, and we cannot determine if
lxc profile remove' failed because it was not present or for
some other reason (outside of string matching).